Handle Deleted Quiz. Logic in Viewing Attempts and Deleting a quiz, improve seeder script accordingly.

+++ src/api/controllers/quizController.js
@@ -1,17 +1,20 @@
-import { Quiz } from '../../models/Quiz.js';
-import { AppError } from '../../utils/AppError.js';
-import { catchAsync } from '../../utils/catchAsync.js';
+import { Attempt } from '../../models/Attempted.js'
+import { Quiz } from '../../models/Quiz.js'
+import { AppError } from '../../utils/AppError.js'
+import { catchAsync } from '../../utils/catchAsync.js'
 
 export const getAllQuizes = catchAsync(async (req, res) => {
-    const {loggedIn} = req.query;
-    const filters={};
-    if(loggedIn) {
-       filters.author = req.user.id;
+    const { loggedIn } = req.query
+    const filters = {
+        deleted: { $ne: true },
+    }
+    if (loggedIn) {
+        filters.author = req.user.id
     } else {
         filters.status = 'active'
     }
-   
-    const quizes = await Quiz.find({ ...filters }).populate('questionsCount');
+
+    const quizes = await Quiz.find({ ...filters }).populate('questionsCount')
 
     return res.status(200).json({
         status: 'success',
@@ -20,23 +23,34 @@ export const getAllQuizes = catchAsync(async (req, res) => {
 })
 
 export const createQuiz = catchAsync(async (req, res, next) => {
-    const { title, description, tags, status } = req.body;
+    const { title, description, tags, status } = req.body
 
-    const author = req.user.id;
+    const author = req.user.id
 
     if (!title || !description || !tags || !author) {
-        return next(new AppError('Please send Quiz title, description, tags and author.', 400))
+        return next(
+            new AppError(
+                'Please send Quiz title, description, tags and author.',
+                400
+            )
+        )
     }
 
-    if(!Array.isArray(tags)) {
-        return next(new AppError('Please send tags as array.', 400))   
+    if (!Array.isArray(tags)) {
+        return next(new AppError('Please send tags as array.', 400))
     }
 
     if (!(tags.length >= 1)) {
         return next(new AppError('Please send at least 1 tag in array.', 400))
     }
 
-    const quiz = await Quiz.create({ title, description, tags, author, status: status });
+    const quiz = await Quiz.create({
+        title,
+        description,
+        tags,
+        author,
+        status: status,
+    })
 
     return res.status(200).json({
         status: 'success',
@@ -45,9 +59,9 @@ export const createQuiz = catchAsync(async (req, res, next) => {
 })
 
 export const getQuiz = catchAsync(async (req, res, next) => {
-    const { quizId } = req.params;
+    const { quizId } = req.params
 
-    const quiz = await Quiz.findById(quizId).populate('questionsCount');
+    const quiz = await Quiz.findById(quizId).populate('questionsCount')
 
     if (!quiz) {
         return next(new AppError('Quiz not found', 404))
@@ -60,28 +74,33 @@ export const getQuiz = catchAsync(async (req, res, next) => {
 })
 
 export const updateQuiz = catchAsync(async (req, res, next) => {
-    const { title, description, tags, status } = req.body;
-    const { quizId } = req.params;
-    const toUpdateData = {};
+    const { title, description, tags, status } = req.body
+    const { quizId } = req.params
+    const toUpdateData = {}
 
     if (title) {
-        toUpdateData.title = title;
+        toUpdateData.title = title
     }
 
     if (description) {
-        toUpdateData.description = description;
+        toUpdateData.description = description
     }
 
     if (tags) {
-        if (!(tags.length >= 1)) return next(new AppError('Please send at least 1 tag.'), 400);
-        toUpdateData.tags = tags;
+        if (!(tags.length >= 1))
+            return next(new AppError('Please send at least 1 tag.'), 400)
+        toUpdateData.tags = tags
     }
 
     if (status) {
-        toUpdateData.status = status;
+        toUpdateData.status = status
     }
 
-    const updatedQuiz = await Quiz.findOneAndUpdate({ _id: quizId }, toUpdateData, { new: true, runValidators: true });
+    const updatedQuiz = await Quiz.findOneAndUpdate(
+        { _id: quizId },
+        toUpdateData,
+        { new: true, runValidators: true }
+    )
 
     return res.status(200).json({
         status: 'success',
@@ -89,11 +108,18 @@ export const updateQuiz = catchAsync(async (req, res, next) => {
     })
 })
 
+export const deleteQuiz = catchAsync(async (req, res) => {
+    const { quizId } = req.params
 
-export const deleteQuiz = catchAsync(async (req, res,) => {
-    const { quizId } = req.params;
+    const attempts = await Attempt.findOne({ quiz: quizId })
 
-    const quiz = await Quiz.findOneAndDelete({ _id: quizId });
+    if (!attempts) {
+        await Quiz.findOneAndDelete({ _id: quizId })
+        return res.status(204).json({
+            status: 'success',
+        })
+    }
+    const quiz = await Quiz.findOneAndUpdate({ _id: quizId }, { deleted: true })
 
     if (!quiz) {
         throw new AppError("Quiz you are trying to delete doesn't exist.", 404)
